# backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Staff Scheduler API",
    "version": "1.0.0",
    "description": "Open-source workforce management REST API.\n\nAll non-public endpoints require a JWT obtained via `POST /api/v1/auth/login`, presented either as the httpOnly `token` cookie (set automatically at login) or as `Authorization: Bearer <token>`.\n\nBase URLs:\n- Canonical: `http://localhost:3001/api/v1`\n- Legacy (will be removed): `http://localhost:3001/api`\n\nEvery response follows the envelope format:\n- Success: `{ \"success\": true, \"data\": <T> }`\n- Error:   `{ \"success\": false, \"error\": { \"code\": \"ERROR_CODE\", \"message\": \"...\" } }`\n\nValidation errors return `VALIDATION_ERROR` with a `details` array of field-level messages.\nRate-limit errors return HTTP 429 with code `RATE_LIMIT_EXCEEDED`.",
    "contact": {
      "name": "Luca Ostinelli",
      "email": "ostinelliluca2@gmail.com",
      "url": "https://github.com/lucaosti/StaffScheduler"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Local development (canonical)"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Local development (legacy — will be removed)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT passed as Authorization: Bearer <token>. Alternative to the httpOnly cookie for non-browser clients."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "httpOnly JWT cookie set by POST /auth/login (primary mechanism for the web frontend)."
      }
    },
    "schemas": {
...
      "Schedule": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "departmentId": {
            "type": "integer"
          },
          "departmentName": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "totalShifts": {
            "type": "integer"
          },
          "totalAssignments": {
            "type": "integer"
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Request body or path parameter failed Zod validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    },
    "parameters": {
...
      "pageQuery": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "limitQuery": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 20
        }
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/schedules": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "summary": "List schedules",
        "parameters": [
          {
            "$ref": "#/components/parameters/pageQuery"
          },
          {
            "$ref": "#/components/parameters/limitQuery"
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "published",
                "archived"
              ]
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Schedule"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Schedules"
        ],
        "summary": "Create schedule",
        "description": "Requires `schedule.manage`. Creates a draft schedule. A schedule cannot overlap with an existing draft or published schedule in the same department.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "startDate": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "endDate": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "departmentId": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  },
                  "templateIds": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "notes": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "startDate",
                  "endDate",
                  "departmentId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Schedule created."
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "Overlapping schedule exists for this department and date range."
          }
        }
      }
    },
    "/schedules/department/{departmentId}": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "summary": "List schedules for a department",
        "parameters": [
          {
            "name": "departmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/schedules/user/{userId}": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "summary": "List schedules involving a user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/schedules/{id}": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "summary": "Get schedule by ID",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule object."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Schedules"
        ],
        "summary": "Update schedule",
        "description": "Updates name, dates, notes, or status. Requires `schedule.manage`. Archived schedules cannot be modified.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "startDate": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "endDate": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "published",
                      "archived"
                    ]
                  },
                  "departmentId": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated."
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Schedules"
        ],
        "summary": "Delete schedule",
        "description": "Only draft schedules can be deleted. Cascades to shifts and assignments. Requires `schedule.manage`.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted."
          },
          "400": {
            "description": "Schedule is not in draft status."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/schedules/{id}/archive": {
      "patch": {
        "tags": [
          "Schedules"
        ],
        "summary": "Archive schedule",
        "description": "Transitions status to `archived`. Requires `schedule.manage`.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Archived."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/schedules/{id}/duplicate": {
      "post": {
        "tags": [
          "Schedules"
        ],
        "summary": "Duplicate schedule",
        "description": "Clones the schedule and all its shifts to a new date range. Requires `schedule.manage`.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "startDate": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "endDate": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  }
                },
                "required": [
                  "name",
                  "startDate",
                  "endDate"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cloned schedule."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/schedules/{id}/generate": {
      "post": {
        "tags": [
          "Schedules"
        ],
        "summary": "Auto-generate assignments",
        "description": "Runs the scheduling optimizer to create pending assignments. Uses the optimal OR-Tools CP-SAT engine by default; if it is unavailable the run degrades to the greedy draft engine and the result reports engine='greedy' with degraded=true (the fallback is always signalled, never silent). Asynchronous by default: when the job queue is enabled (Redis) it returns 202 with a job id — poll GET /schedules/{id}/optimization for progress/result. Without Redis it runs synchronously and returns 200 with the result. Requires schedule.optimize.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Generation result.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "assignmentsCreated": {
                      "type": "integer"
                    },
                    "totalShifts": {
                      "type": "integer"
                    },
                    "coveragePercentage": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string"
                    },
                    "engine": {
                      "type": "string",
                      "enum": [
                        "or-tools",
                        "greedy"
                      ],
                      "description": "Engine that produced the schedule: 'or-tools' is the optimal CP-SAT solver, 'greedy' is the best-effort draft engine."
                    },
                    "degraded": {
                      "type": "boolean",
                      "description": "True when the optimal engine was requested but the run fell back to greedy (e.g. OR-Tools unavailable). The result is a draft, not the optimum."
                    },
                    "degradedReason": {
                      "type": "string",
                      "description": "Why the run degraded to greedy, when it did."
                    }
                  }
                }
              }
            }
          },
          "202": {
            "description": "Optimization queued; poll the optimization endpoint for status.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobId": {
                          "type": "string"
                        },
                        "scheduleId": {
                          "type": "integer"
                        },
                        "state": {
                          "type": "string"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/schedules/{id}/publish": {
      "patch": {
        "tags": [
          "Schedules"
        ],
        "summary": "Publish schedule",
        "description": "Transitions status from `draft` to `published`. The schedule must have at least one shift. Requires `schedule.publish`.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Published."
          },
          "400": {
            "description": "No shifts or wrong status."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/schedules/{id}/shifts": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "summary": "List shifts in schedule",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Shift list with staffing counts."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/shifts/schedule/{scheduleId}": {
      "get": {
        "tags": [
          "Shifts"
        ],
        "summary": "List shifts of a schedule",
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/schedules/{id}/optimization": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "summary": "Optimization job status",
        "description": "Returns the state, progress and result of the schedule's optimization job. Requires schedule.optimize.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Job status.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobId": {
                          "type": "string"
                        },
                        "state": {
                          "type": "string",
                          "enum": [
                            "waiting",
                            "active",
                            "completed",
                            "failed",
                            "unknown"
                          ]
                        },
                        "progress": {
                          "type": "number"
                        },
                        "result": {
                          "type": "object",
                          "properties": {
                            "assignmentsCreated": {
                              "type": "integer"
                            },
                            "totalShifts": {
                              "type": "integer"
                            },
                            "coveragePercentage": {
                              "type": "number"
                            },
                            "status": {
                              "type": "string"
                            },
                            "engine": {
                              "type": "string",
                              "enum": [
                                "or-tools",
                                "greedy"
                              ],
                              "description": "Engine that produced the schedule."
                            },
                            "degraded": {
                              "type": "boolean",
                              "description": "True when the optimum was requested but the run fell back to greedy."
                            },
                            "degradedReason": {
                              "type": "string"
                            }
                          }
                        },
                        "failedReason": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Schedules"
        ],
        "summary": "Cancel optimization job",
        "description": "Cancels the schedule's in-flight or retained optimization job. Requires schedule.optimize.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  }
}